Fixed compiler warnings.

+++ stasis/common.h
@@ -100,9 +100,10 @@ extern int errno;
 typedef unsigned char byte;
 //@todo lsn_t should be unsigned.
 // If it were unsigned, it could be typedef'ed from size_t.
-typedef int64_t lsn_t;
+// Setting these to int64_t makes printf cranky. :(
+typedef long long lsn_t;
 #define LSN_T_MAX    INT64_MAX
-typedef int64_t pageid_t;
+typedef long long pageid_t;
 #define PAGEID_T_MAX INT64_MAX
 
 /*#define DEBUGGING   */
